finanzas_app/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
firebase-admin>=6.2.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dateutil
google-api-python-client
google-auth

# Producción
gunicorn
whitenoise
dj-database-url
python-dotenv

# Testing
pytest
pytest-django
pytest-cov
factory-boy
